+++ tests/test_error.py
@@ -36,7 +36,9 @@ def test_critical_errors(caplog):
     stage = TextReverser()
     manager = ErrorManager()
     item = DataItem()
-    managed_critical_error = manager.handle(CriticalError(), stage, item)
+    error = CriticalError()
+    error.with_exception(Exception())
+    managed_critical_error = manager.handle(error, stage, item)
     assert not item.has_errors()
     assert item.has_critical_errors()
     assert isinstance(
@@ -46,8 +48,11 @@ def test_critical_errors(caplog):
     assert any(caplog.records)
     manager = ErrorManager().raise_on_critical_error()
     item = DataItem()
-    with pytest.raises(Exception):
+    with pytest.raises(CriticalError):
         manager.handle(CriticalError(), stage, item)
+    with pytest.raises(Exception):
+        error = CriticalError().with_exception(Exception())
+        manager.handle(error, stage, item)
     assert any(caplog.records)
     assert item.has_critical_errors()
     assert not item.has_errors()