+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - BREAKING CHANGES:
   - `use_threads` parameter has been substituted with `parallel`, values are inverted
   - item timings are now in seconds
+  - `get_exception` of an Error without associated Exception now returns `None`
 
 ## Version 0.3.0 [2020-04-22]